feat: added context for wasm components
