+++ Cargo.toml
@@ -75,6 +75,7 @@ bytes = "1.4"
 base64 = "0.21"
 bitflags = "2.2"
 cfg-if = "1.0"
+check_keyword = "0.2"
 chrono = "0.4"
 clap = "4.1"
 console = "0.15"
