+++ crates/wick/flow-graph-interpreter/tests/manifests/v1/inline-node-ids.yaml
@@ -7,14 +7,6 @@ component:
   kind: wick/component/composite@v1
   operations:
     - name: testop
-      inputs:
-        - name: message
-          type: string
-        - name: component
-          type: link
-      outputs:
-        - name: output
-          type: string
       flow:
         - <> -> test::reverse[A].input
         - A.output -> test::uppercase[B].input
