+++ crates/wick/wick-config/Cargo.toml
@@ -31,6 +31,7 @@ url = { workspace = true }
 glob = { workspace = true }
 parking_lot = { workspace = true }
 liquid-json = { workspace = true, features = ["serde"] }
+itertools = { workspace = true }
 
 [dev-dependencies]
 wick-xdg = { workspace = true }
