+++ crates/wick/wick-config/docs/v1.html
@@ -1888,6 +1888,20 @@
 </div><div class=field>
 
 <h3 class=field-name>
+  <a name="with">with</a>
+</h3>
+
+<span class="description field-description">The configuration for the operation, if any.</span>
+
+
+<ul class="field-notes">
+<li><span class="type field-note">Type: <span class=field-type>Map<string,<a href="#any">any</a>><span class=optional>(Optional)</span></span></span></li>
+
+
+</ul>
+</div><div class=field>
+
+<h3 class=field-name>
   <a name="input">input</a>
 </h3>
 
@@ -2036,7 +2050,7 @@
 </div><div class=field>
 
 <h3 class=field-name>
-  <a name="message">message</a>
+  <a name="error">error</a>
 </h3>
 
 <span class="description field-description">The error message.</span>
