+++ crates/wick/wick-config/json-schema/v1/manifest.json
@@ -1269,18 +1269,23 @@
         },
         "operation": {
           "description": "The operaton to test.",
-          "oneOf": [
-            { "type": "string" },
-            {
-              "type": "string"
-            }
-          ]
+
+          "type": "string"
         },
         "inherent": {
           "description": "Inherent data to use for the test.",
           "required": false,
 
           "$ref": "#/$defs/v1/InherentData"
+        },
+        "with": {
+          "description": "The configuration for the operation, if any.",
+          "required": false,
+
+          "type": "object",
+          "patternProperties": {
+            "[a-zA-Z0-9][a-zA-Z0-9_]*": {}
+          }
         },
         "input": {
           "description": "The inputs to the test.",
@@ -1370,13 +1375,13 @@
 
           "$ref": "#/$defs/v1/PacketFlags"
         },
-        "message": {
+        "error": {
           "description": "The error message.",
 
           "type": "string"
         }
       },
-      "required": ["name", "message"]
+      "required": ["name", "error"]
     },
 
     "PacketFlags": {
