+++ crates/wick/wick-host/src/collection.rs
@@ -47,10 +47,10 @@ impl Component for HostComponent {
     &self,
     invocation: Invocation,
     stream: PacketStream,
-    _data: Option<wick_packet::OperationConfig>,
+    data: Option<wick_packet::OperationConfig>,
     _callback: Arc<RuntimeCallback>,
   ) -> flow_component::BoxFuture<Result<PacketStream, ComponentError>> {
-    let fut = self.host.invoke(invocation, stream);
+    let fut = self.host.invoke(invocation, stream, data);
 
     Box::pin(async move {
       let outputs = fut.await.map_err(ComponentError::new)?;
