+++ tests/testdata/files/tests.wick
@@ -4,11 +4,8 @@ tests:
     operation: test-component
     input:
       - name: input
-        message: 'test'
+        error: 'test'
       - name: input
         data: 2
       - name: input
         data: hey
-    # outputs:
-    #   - port: output
-    #     data: test
