+++ tests/testdata/manifests/configuration.wick
@@ -1,0 +1,36 @@
+---
+name: 'test'
+kind: wick/component@v1
+import:
+  - name: test_component
+    component:
+      kind: wick/component/manifest@v1
+      ref: ../../../crates/integration/test-baseline-component/component.yaml
+metadata:
+  version: '0.0.2'
+component:
+  kind: wick/component/composite@v1
+  operations:
+    - name: test
+      uses:
+        - name: test_component
+          operation: test_component::power
+          with:
+            exponent: 4
+      flow:
+        - <> -> test_component.input
+        - test_component.output -> <>
+tests:
+  - name: test
+    operation: test
+    with:
+      exponent: 6
+    input:
+      - name: input
+        data: 2
+    output:
+      - name: output
+        data: 16
+      - name: output
+        flags:
+          done: true
